README updated with a quick start
First off thanks a lot for the awesome library :).  README was updated with quick start steps on how to integrate the library with an Ionic app. 

Having a quick start will allow developers to easily integrated in with their `ionic` apps. 

If you need any changes/ alternations , I'm happy to do them and send it back.

Keep up the good work

cheers